3d works, also changed everything to double precision

+++ src/functions.h
@@ -8,7 +8,7 @@
 	k=1 (uni-variate)
 	params = []  (no parameters)
 */
-__device__ float F0(const float *x, const float *params)
+__device__ double F0(const double *x, const double *params)
 { return exp(x[0]); }
 
 /* Name = sin(o1+x*y)*exp(o2+x)
@@ -16,7 +16,7 @@ __device__ float F0(const float *x, const float *params)
 	k=2 (bi-variate)
 	params = [o1,o2]  (2 parameters)
 */
-__device__ float F1(const float *x, const float *params)
+__device__ double F1(const double *x, const double *params)
 {
 	return sin(params[0] + x[0]*x[1]) * exp(params[1]+x[0]);
 }
@@ -26,7 +26,7 @@ __device__ float F1(const float *x, const float *params)
 	k=3 (tri-variate)
 	params = []  (no parameters)
 */
-__device__ float F2(const float *x, const float *params)
+__device__ double F2(const double *x, const double *params)
 {
 	return round(exp(-x[0]))-round(exp(x[1]))*sin(x[2]);
 }
@@ -36,10 +36,10 @@ __device__ float F2(const float *x, const float *params)
 	k=3 (tri-variate)
 	params = [norm_lev]  (1 parameter)
 */
-__device__ float F3(const float *x, const float *params)
+__device__ double F3(const double *x, const double *params)
 {
 	// Perform a vector norm of a specific power, e.g. if params[0]==2 then it is the euclidian norm
-	float v=pow(x[0],params[0]) + pow(x[1],params[0]) + pow(x[2],params[0]);
+	double v=pow(x[0],params[0]) + pow(x[1],params[0]) + pow(x[2],params[0]);
 	v=pow(v,1/params[0]);
 	// Then a sinc
 	return sin(v)/v;
@@ -56,15 +56,15 @@ __device__ float F3(const float *x, const float *params)
 	params. Don't worry too much about the details of the
 	matrix, it isn't relative to making it go fast.
 */
-__device__ float F4(const float *x, const float *params)
+__device__ double F4(const double *x, const double *params)
 {
 	// We need to calculate scale*exp(-x'*\Sigma^{-1}*x/2),
 	// where \Sigma^{-1} is a 3x3 matrix.
 	int i;
-	float acc=0;
+	double acc=0;
 	for(i=0;i<3;i++){
 		// We first form a row of the \Sigma^-1 * x
-		float row= params[i*3+0]*x[0] + params[i*3+1]*x[1] + params[i*3+2]*x[2];
+		double row= params[i*3+0]*x[0] + params[i*3+1]*x[1] + params[i*3+2]*x[2];
 		// Then perform the left mutiply with x'
 		acc += row*x[i];
 	}
@@ -76,7 +76,7 @@ __device__ float F4(const float *x, const float *params)
 	k=3 (tri-variate)
 	params= [] (no parameters)
 */
-__device__ float F5(const float *x, const float *params)
+__device__ double F5(const double *x, const double *params)
 {
 	return powf(powf(sinf( pow(x[0], x[1]) ), 2), x[2]);
 }
@@ -86,10 +86,10 @@ __device__ float F5(const float *x, const float *params)
 	k=3 (tri-variate)
 	params= [radius,width]
 */
-__device__ float F6(const float *x, const float *params)
+__device__ double F6(const double *x, const double *params)
 {
-	float a=powf(powf(x[0],2)+powf(x[1],2)+powf(x[2],2),0.5); // distance from origin
-	float d=(a-params[0]);	// How far from surface of sphere
+	double a=powf(powf(x[0],2)+powf(x[1],2)+powf(x[2],2),0.5); // distance from origin
+	double d=(a-params[0]);	// How far from surface of sphere
 	if(d<0)
 		d=-d;
 	if(d<params[1]){
